Give audio tests it can fail, and fix the click the first one found
Plan E2's selftest half. 29 audio cases now drive the mixer directly and
check exact values, and every expectation is derived from the sample
format's own arithmetic rather than taken from what our mixer currently
produces. That distinction matters here: this project has already blessed
a frame-hash manifest from broken output and spent a day with the bug
pinned in place by its own test.

One case failed on its first run and the failure was real. The resampler
formed the difference between two samples times a fraction in a 32-bit
int, and a difference near full scale times a fraction near one overflows
by exactly 32,768, wrapping a silent sample to full-scale negative. That
is an audible click on any loud sample that crosses near full scale
between adjacent source samples, and the file already disagreed with
itself, because the mixing command does the same multiply in a wider
type. Widened.

The port also speaks, which nobody had established. The new voice
scenario reaches the end of the first battle and the run opens a
three-digit voice line. The opening genuinely contains no speech: of 258
field scripts, 66 name a stream and every one of those names only the
low-frequency ambience numbers, and the opening's own script names
exactly one of them. So the zero voice-bank count that the census
reported was never a driver defect.

Two further corrections to the census, both worth knowing before anyone
trusts its output again. The bank that showed zero starts is a
sound-effect bank, not a voice bank; its waveform ids are disjoint from
the other effect bank and together they partition one id space. And the
census cannot name a voice bank at all: it identifies an upload by a
32-byte prefix or by its size, and the 72 speaker banks come in identical
twins, two of which match across all 262,448 bytes, while the voice
streams have byte-identical left and right halves. Across the disc, 23%
of files can be named by neither test.

+++ runtime/ax.c
@@ -501,7 +501,12 @@ static void process_voice(CpuState* s, uint32_t addr, uint8_t* aram)
                 last[0] = last[1]; last[1] = last[2]; last[2] = last[3];
                 last[3] = fetch_sample(&v);
             }
-            samples[i] = last[2] + (((last[3] - last[2]) * (int32_t)frac) >> 16);
+            /* In int, a difference near full scale times a fraction near one
+             * overflows: 32769 * 65535 is 32,768 past the maximum and wraps
+             * to full-scale negative, turning a silent sample into a click.
+             * Command 0x01's mix already does its multiply wide; this is the
+             * same arithmetic and needs the same width. */
+            samples[i] = last[2] + (int32_t)(((int64_t)(last[3] - last[2]) * (int64_t)frac) >> 16);
         }
 
         /* volume envelope */
